Big Additions: Elastic API and Controller Creation Handling
Added the Elastic.java file which contains methods to interact with Elastic Dashboard.

Also added SKController.java, which allows controller construction to be made simple by specifying the type and port in its constructor. It will handle the binding of buttons no matter the type to the same collection of methods (which create new SKTrigger objects).

Lastly, added SKCommandXboxController which extends SKTrigger and makes new triggers for FilteredXboxController objects.